steamshim: make RequestAvatar work for non friends
