+++ source/steamshim/src/child/child.cpp
@@ -65,6 +65,7 @@ static STEAMSHIM_Event* ProcessEvent(){
     PRINTGOTEVENT(SHIMEVENT_PERSONANAMERECIEVED);
     PRINTGOTEVENT(SHIMEVENT_AUTHSESSIONTICKETRECIEVED);
     PRINTGOTEVENT(SHIMEVENT_AUTHSESSIONVALIDATED);
+    PRINTGOTEVENT(SHIMEVENT_AVATARRECIEVED);
     #undef PRINTGOTEVENT
     else printf("Child got unknown shimevent %d.\n", (int) type);
     #endif
